+++ src/pages/accountRecovery/findAccountPage.tsx
@@ -1,45 +1,16 @@
 import { Link, useNavigate } from 'react-router-dom';
-import { useState } from 'react';
-import { useFindAccountMutation } from '@/api/auth/account/accountHooks';
 import { logo_sm } from '@/assets/assets';
 import Button from '@/components/common/Button';
 import Spinner from '@/components/common/Spinner';
 import { useTranslation } from 'react-i18next';
 import PhoneInput from '@/components/domain/auth/PhoneInput';
+import { useFindAccount } from './useFindAccount';
 
 const FindAccountPage = () => {
   const navigate = useNavigate();
   const { t } = useTranslation();
-  const [phoneNumber, setPhoneNumber] = useState('');
-  const [fullPhoneNumber, setFullPhoneNumber] = useState('');
-  const [error, setError] = useState('');
-  const [foundAccounts, setFoundAccounts] = useState<any[]>([]);
-
-  const findAccountMutation = useFindAccountMutation({
-    onSuccess: (data) => {
-      if (data.accounts.length > 0) {
-        setFoundAccounts(data.accounts);
-        setError('');
-      } else {
-        setError('등록된 계정이 없습니다.');
-      }
-    },
-    onError: (error: any) => {
-      setError(error.response?.data?.message || '계정 찾기에 실패했습니다.');
-    },
-  });
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-
-    if (!phoneNumber) {
-      setError(t('auth.please_enter_phone'));
-      return;
-    }
-
-    setError('');
-    findAccountMutation.mutate({ phone_number: fullPhoneNumber });
-  };
+  const { formState, handlers, isLoading } = useFindAccount();
+  const { phoneNumber, error, foundAccounts } = formState;
 
   return (
     <div className='flex min-h-screen items-center justify-center sm:px-6 lg:px-8'>
@@ -85,37 +56,20 @@ const FindAccountPage = () => {
               </p>
             </div>
 
-            <form className='mt-8 space-y-6' onSubmit={handleSubmit}>
+            <form className='mt-8 space-y-6' onSubmit={handlers.handleSubmit}>
               <div>
                 <PhoneInput
                   label={t('auth.phone_number')}
-                  placeholder={t('auth.phone_number_placeholder')}
                   value={phoneNumber}
-                  onChange={(cleanValue, fullNumber) => {
-                    setPhoneNumber(cleanValue);
-                    setFullPhoneNumber(fullNumber);
-                    setError('');
-                  }}
-                  onClear={() => {
-                    setPhoneNumber('');
-                    setFullPhoneNumber('');
-                    setError('');
-                  }}
+                  onChange={handlers.handlePhoneChange}
+                  onClear={handlers.handleReset}
                   defaultCountry='KR'
                 />
                 {error && <p className='mt-2 text-sm text-red-600'>{error}</p>}
               </div>
 
-              <Button
-                type='submit'
-                disabled={findAccountMutation.isPending}
-                className='w-full'
-              >
-                {findAccountMutation.isPending ? (
-                  <Spinner />
-                ) : (
-                  t('auth.find_email')
-                )}
+              <Button type='submit' disabled={isLoading} className='w-full'>
+                {isLoading ? <Spinner /> : t('auth.find_email')}
               </Button>
             </form>
